refactor: extract search datafusion optimizers
