+++ src/core/src/service/search/datafusion/optimizer/physical_optimizer/mod.rs
@@ -23,4 +23,4 @@ pub mod index_optimizer;
 pub mod join_reorder;
 pub mod remote_scan;
 pub mod rewrite_match;
-pub mod utils;
+pub use search::datafusion::optimizer::physical_optimizer::utils;
