+++ src/search/src/datafusion/distributed_plan/empty_exec.rs
