+++ src/search/src/datafusion/mod.rs
@@ -13,8 +13,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+pub mod distributed_plan;
+pub mod optimizer;
 pub mod peak_memory_pool;
 pub mod plan;
 pub mod planner;
+pub mod table_provider;
 pub mod udaf;
 pub mod udf;
