+++ src/search/src/datafusion/optimizer/analyze/mod.rs
