+++ src/search/src/datafusion/optimizer/analyze/remove_index_fields.rs
