+++ src/search/src/datafusion/optimizer/physical_optimizer/utils.rs
